Separate city simulation code from drawing + add Wojtek's suggestions
